# backend/src/public/swagger.yaml
openapi: 3.0.0
info:
  title: Forkcast API
  description: API Documentation for Forkcast Application
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PublicUser:
      type: object
      required:
        - id
        - name
        - profilePicture
      properties:
        id:
          type: integer
        name:
          type: string
        profilePicture:
          type: string
          format: uri
          nullable: true
        isVerified:
           type: boolean
...
    SendFriendRequest:
      type: object
      required:
        - username
      properties:
        username:
          type: string
    SendFriendRequestResult:
      type: object
      properties:
        status:
          type: string
          enum: [pending, accepted]
        user:
          $ref: '#/components/schemas/PublicUser'
    IncomingFriendRequest:
      allOf:
        - $ref: '#/components/schemas/PublicUser'
        - type: object
          properties:
            requestId:
              type: integer
...
security:
  - bearerAuth: []
paths:
...
  /users/me/friends:
    get:
      summary: Get friend list
      tags: [Friends]
      responses:
        '200':
          description: Friend list mappings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicUser'
        '401':
          description: Missing or invalid authentication
        '500':
          description: Internal server error
  /users/me/friends/requests:
    get:
      summary: Get incoming (non-expired) friend requests
      tags: [Friends]
      responses:
        '200':
          description: Incoming friend requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IncomingFriendRequest'
        '401':
          description: Missing or invalid authentication
        '500':
          description: Internal server error
    post:
      summary: Send a friend request by username
      tags: [Friends]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendFriendRequest'
      responses:
        '201':
          description: Friend request sent (pending)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendFriendRequestResult'
        '200':
          description: Auto-accepted because the target had already requested you
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendFriendRequestResult'
        '400':
          description: Invalid request body or adding self
        '404':
          description: Username not found
        '409':
          description: Already friends or request already sent
        '401':
          description: Missing or invalid authentication
        '500':
          description: Internal server error
  /users/me/friends/requests/{id}/accept:
    post:
      summary: Accept an incoming friend request
      tags: [Friends]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Request accepted; returns the new friend's profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUser'
        '404':
          description: Friend request not found
        '401':
          description: Missing or invalid authentication
        '500':
          description: Internal server error
  /users/me/friends/requests/{id}/decline:
    post:
      summary: Decline an incoming friend request
      tags: [Friends]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Request declined
        '404':
          description: Friend request not found
        '401':
          description: Missing or invalid authentication
        '500':
          description: Internal server error
  /users/me/friends/{friendId}:
    get:
      summary: Get a specific friend's profile
      tags: [Friends]
      parameters:
        - in: path
          name: friendId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Friend profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUser'
        '400':
          description: Invalid friend ID
        '401':
          description: Missing or invalid authentication
        '404':
          description: Friend or user not found
        '500':
          description: Internal server error
    delete:
      summary: Remove a friend
      tags: [Friends]
      parameters:
        - in: path
          name: friendId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Friend removed
        '400':
          description: Invalid friend ID
        '401':
          description: Missing or invalid authentication
        '404':
          description: Friend connection not found
        '500':
          description: Internal server error
...